add first draft of get api for scores
