+++ saguaro_tom/api.py
@@ -1,0 +1,5 @@
+from ninja import NinjaAPI
+
+api = NinjaAPI()
+
+api.add_router("/score/", "candidate_vetting.api.router")
